Add make_safe to ask_cfpb
